+++ package.js
@@ -23,15 +23,15 @@ Package.onTest( function( api ){
 });
 
 function configure( api ){
-    api.versionsFrom( '1.8.1' );
+    api.versionsFrom( '1.13.2' );
     api.use( 'alanning:roles@3.4.0', 'server' );
     api.use( 'blaze-html-templates@2.0.0', 'client' );
     api.use( 'ecmascript' );
     api.use( 'less@4.0.0', 'client' );
     api.use( 'mongo', 'server' );
-    api.use( 'pwix:i18n@1.3.0' );
+    api.use( 'pwix:i18n@1.5.2' );
     api.use( 'pwix:jstree@1.0.3' );
-    api.use( 'pwix:modal@1.5.0' );
+    api.use( 'pwix:modal@1.7.1' );
     api.use( 'tmeasday:check-npm-versions@1.0.2', 'server' );
     api.addFiles( 'src/client/components/prEdit/prEdit.js', 'client' );
     api.addFiles( 'src/client/components/prView/prView.js', 'client' );